final fixes for main processes

+++ app/handlers/users/appeal.py
@@ -1,6 +1,7 @@
 import asyncio
 from contextlib import suppress
 from datetime import datetime
+from typing import Union
 
 from aiogram import Router, F, Bot, types
 from aiogram.exceptions import TelegramBadRequest, TelegramRetryAfter
@@ -26,13 +27,21 @@ router.callback_query.filter(F.message.chat.type == 'private', CommonFilter())
 
 @router.callback_query(CheckingCallback.filter(F.param == "appeal"), F.message.chat.type == 'private')
 async def start_appeal(call: types.CallbackQuery, state: FSMContext, callback_data: CheckingCallback,
-                       db_session: sessionmaker, user: User):
-    appeal = await validate_appeal(db_session, user, callback_data.ticket_id)
-
+                       db_session: sessionmaker):
+    appeal: Union[Ticket, str] = await validate_appeal(db_session, callback_data.ticket_id)
+    if isinstance(appeal, str):
+        with suppress(TelegramBadRequest):
+            await call.message.edit_text(call.message.html_text + '\n\n' + appeal, reply_markup=None)
+        return
     await call.message.delete()
-    await call.message.answer('Пожалуйста, напиши обоснование к апелляции (не более 4000 символов).',
+    await call.message.answer(f'{"Апелляция" if appeal.status_id == 4 else "Кассация"} по клиенту:\n'
+                              f'<a href="{appeal.link}">{appeal.fullname}</a>\n\n'
+                              f'Пожалуйста, напиши обоснование к апелляции (не более 4000 символов).',
                               reply_markup=ForceReply(input_field_placeholder='Текст апелляции'))
-    await state.update_data(ticket_id=callback_data.ticket_id)
+    await state.update_data(
+        ticket_id=callback_data.ticket_id,
+        new_status=5 if appeal.status_id == 4 else 6
+    )
     await state.set_state(Appeal.comment)
 
 
@@ -43,18 +52,22 @@ async def send_appeal(msg: types.Message, state: FSMContext, user: User, db_sess
         await msg.answer('Я принимаю только текстовые сообщения, без файлов, фотографий и прочего.')
         return
     state_data = await state.get_data()
-    ticket_id = state_data['ticket_id']
+    ticket_id, new_status_id = list(map(int, state_data.values()))
+
     appeal_text = msg.text
     async with db_session() as session:
         try:
-            ticket = TicketHistory(
-                ticket_id=int(ticket_id),
+            ticket = await session.get(Ticket, ticket_id)
+            session.add(TicketHistory(
+                ticket_id=ticket_id,
                 sender_id=msg.from_user.id,
-                status_id=5,
+                status_id=new_status_id,
                 comment=appeal_text
-            )
-            session.add(ticket)
-            await session.merge(Ticket(id=int(ticket_id), status_id=5, comment=msg.text, updated_at=func.now()))
+            ))
+            ticket.status_id = new_status_id
+            ticket.comment = appeal_text
+            ticket.updated_at = func.now()
+
             await session.commit()
         except Exception as e:
             logger.error(e)
@@ -63,18 +76,20 @@ async def send_appeal(msg: types.Message, state: FSMContext, user: User, db_sess
             return
     successful = False
     await state.clear()
+    appeal_type = "🟡 Апелляция" if new_status_id == 5 else "🔴 Кассация"
     while not successful:
         try:
             await bot.send_message(
                 chat_id=config.misc.checking_group,
-                text=f'🔴 <b>Апелляция по клиенту</b>\n'
-                     f'{ticket.link}\n\n'
-                     f'<u>Кто подал:</u>\n{user.fullname} @{msg.from_user.username}\n'
-                     f'Когда подал: <b>{datetime.now().strftime("%d.%m.%Y %H:%M:%S")}</b>\n\n'
+                text=f' <b>{appeal_type} по клиенту</b>:\n'
+                     f'<b><a href="{ticket.link}">{ticket.fullname}</a></b>\n\n'
+                     f'<u>Автор:</u>\n{user.fullname} @{msg.from_user.username}\n'
+                     f'Поступление: <b>{datetime.now().strftime("%d.%m.%Y %H:%M:%S")}</b>\n\n'
                      f'<i>Обоснование:</i>\n{appeal_text}',
                 reply_markup=await get_check_keyboard(ticket_id, user.id)
             )
-            logger.opt(lazy=True).log('APPEAL', f'User {user.fullname} successfully sent client (ID: {ticket.id})')
+            logger.opt(lazy=True).log('APPEAL', f'User {user.fullname} sent appeal for client (type: {appeal_type} |'
+                                                f'ID: {ticket.id})')
             successful = True
         except TelegramRetryAfter as e:
             await asyncio.sleep(e.retry_after)
